keydb.conf: support globbing in include directive
The existing example keydb.conf indicates that wildcard pattern matching
is supported for the include directive, i.e. it implies that:

- include /etc/keydb.conf
- include /etc/keydb.d/server-specific.conf
- include /etc/keydb.d/*.conf

should all be supported. However, glob-style matching support is not
enabled by the configuration parser. This commit treats the arguments to
the include directive as glob patterns via the POSIX standard glob.h
library, and calls the loadServerConfig function as many times as
necessary.

Signed-off-by: Blake Alexander <blake.alexander@airbus.com>